refactor(frontend): modularize app shell and establish centralized design system foundation
