+++ frontend/src/components/TeacherLoginPage.jsx
@@ -1,5 +1,7 @@
 import { useState } from 'react'
 import { useAuth } from '../contexts/AuthContext'
+import { TEACHER_LOGIN_COPY } from '../content/strings'
+import { AppShell, Button, Input, Panel } from './ui/primitives'
 
 export default function TeacherLoginPage({ onSuccess }) {
   const { login, register, loginWithGoogle } = useAuth()
@@ -42,10 +44,10 @@ export default function TeacherLoginPage({ onSuccess }) {
   }
 
   return (
-    <div className="min-h-screen bg-gray-50 flex flex-col items-center justify-center px-4">
+    <AppShell className="flex flex-col items-center justify-center px-4">
       <div className="w-full max-w-sm">
         <div className="text-center mb-8">
-          <div className="w-14 h-14 rounded-2xl bg-indigo-50 border border-indigo-100 flex items-center justify-center mx-auto mb-3">
+          <div className="w-14 h-14 rounded-2xl bg-[var(--color-brand-50)] border border-[var(--color-brand-100)] flex items-center justify-center mx-auto mb-3">
             <svg xmlns="http://www.w3.org/2000/svg" className="w-8 h-8 text-indigo-500" viewBox="0 0 24 24" fill="currentColor">
               <ellipse cx="5.5" cy="6.5" rx="2.5" ry="3" />
               <ellipse cx="10" cy="4" rx="2" ry="2.5" />
@@ -54,11 +56,11 @@ export default function TeacherLoginPage({ onSuccess }) {
               <path d="M12 9c-4.418 0-8 2.686-8 6 0 2.21 3.582 4 8 4s8-1.79 8-4c0-3.314-3.582-6-8-6z" />
             </svg>
           </div>
-          <h2 className="text-2xl font-bold text-gray-800">Teacher Portal</h2>
-          <p className="text-gray-400 text-sm mt-1">Sign in to manage your modules</p>
+          <h2 className="text-2xl font-bold text-gray-800">{TEACHER_LOGIN_COPY.title}</h2>
+          <p className="text-gray-400 text-sm mt-1">{TEACHER_LOGIN_COPY.subtitle}</p>
         </div>
 
-        <div className="bg-white rounded-2xl shadow-sm border border-gray-200 p-6">
+        <Panel className="p-6">
           {/* mode toggle */}
           <div className="flex rounded-lg bg-gray-100 p-1 mb-5">
             <button
@@ -68,7 +70,7 @@ export default function TeacherLoginPage({ onSuccess }) {
                 mode === 'signin' ? 'bg-white text-gray-800 shadow-sm' : 'text-gray-500 hover:text-gray-700',
               ].join(' ')}
             >
-              Sign In
+              {TEACHER_LOGIN_COPY.signIn}
             </button>
             <button
               onClick={() => { setMode('register'); setError('') }}
@@ -77,7 +79,7 @@ export default function TeacherLoginPage({ onSuccess }) {
                 mode === 'register' ? 'bg-white text-gray-800 shadow-sm' : 'text-gray-500 hover:text-gray-700',
               ].join(' ')}
             >
-              Create Account
+              {TEACHER_LOGIN_COPY.createAccount}
             </button>
           </div>
 
@@ -89,54 +91,55 @@ export default function TeacherLoginPage({ onSuccess }) {
             )}
 
             {mode === 'register' && (
-              <input
+              <Input
                 type="text"
                 value={displayName}
                 onChange={(e) => setDisplayName(e.target.value)}
-                placeholder="Your name"
-                className="w-full rounded-lg border border-gray-200 px-3 py-2.5 text-sm focus:outline-none focus:ring-2 focus:ring-indigo-400 focus:border-transparent"
+                placeholder={TEACHER_LOGIN_COPY.displayNamePlaceholder}
               />
             )}
 
-            <input
+            <Input
               type="email"
               value={email}
               onChange={(e) => setEmail(e.target.value)}
-              placeholder="Email address"
+              placeholder={TEACHER_LOGIN_COPY.emailPlaceholder}
               required
               autoComplete="email"
-              className="w-full rounded-lg border border-gray-200 px-3 py-2.5 text-sm focus:outline-none focus:ring-2 focus:ring-indigo-400 focus:border-transparent"
             />
 
-            <input
+            <Input
               type="password"
               value={password}
               onChange={(e) => setPassword(e.target.value)}
-              placeholder="Password"
+              placeholder={TEACHER_LOGIN_COPY.passwordPlaceholder}
               required
               autoComplete={mode === 'signin' ? 'current-password' : 'new-password'}
-              className="w-full rounded-lg border border-gray-200 px-3 py-2.5 text-sm focus:outline-none focus:ring-2 focus:ring-indigo-400 focus:border-transparent"
             />
 
-            <button
+            <Button
               type="submit"
               disabled={loading || !email || !password}
-              className="w-full rounded-lg bg-indigo-600 text-white text-sm font-semibold py-2.5 hover:bg-indigo-700 disabled:opacity-40 transition-colors mt-1"
+              variant="primary"
+              size="lg"
+              className="w-full mt-1"
             >
-              {loading ? 'Please wait...' : mode === 'signin' ? 'Sign In' : 'Create Account'}
-            </button>
+              {loading ? TEACHER_LOGIN_COPY.wait : mode === 'signin' ? TEACHER_LOGIN_COPY.signIn : TEACHER_LOGIN_COPY.createAccount}
+            </Button>
           </form>
 
           <div className="flex items-center gap-3 my-4">
             <div className="flex-1 h-px bg-gray-200" />
-            <span className="text-xs text-gray-400">or</span>
+            <span className="text-xs text-gray-400">{TEACHER_LOGIN_COPY.or}</span>
             <div className="flex-1 h-px bg-gray-200" />
           </div>
 
-          <button
+          <Button
             onClick={handleGoogle}
             disabled={loading}
-            className="w-full rounded-lg border border-gray-200 bg-white text-gray-700 text-sm font-medium py-2.5 flex items-center justify-center gap-2 hover:bg-gray-50 disabled:opacity-40 transition-all"
+            variant="secondary"
+            size="lg"
+            className="w-full flex items-center justify-center gap-2"
           >
             <svg className="w-4 h-4" viewBox="0 0 24 24">
               <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
@@ -144,11 +147,11 @@ export default function TeacherLoginPage({ onSuccess }) {
               <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
               <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
             </svg>
-            Sign in with Google
-          </button>
-        </div>
+            {TEACHER_LOGIN_COPY.googleSignIn}
+          </Button>
+        </Panel>
       </div>
-    </div>
+    </AppShell>
   )
 }
