+++ frontend/src/content/strings.js
@@ -1,0 +1,98 @@
+export const APP_COPY = {
+  appName: "Teacher's Pet",
+  appTagline: 'AI-Powered Math Tutor',
+  loading: 'Loading...',
+  leave: 'Leave',
+  back: '← Back',
+}
+
+export const ENTRY_COPY = {
+  studentTitle: "I'm a Student",
+  studentSubtitle: 'Enter your course code to start learning',
+  teacherTitle: "I'm a Teacher",
+  teacherSubtitle: 'Manage modules and view analytics',
+  guestTitle: 'Just Chat',
+  guestSubtitle: 'Continue without signing in',
+}
+
+export const STUDENT_ENTRY_COPY = {
+  title: 'Join a Class',
+  subtitle: 'Enter the code your teacher gave you',
+  inputPlaceholder: 'e.g. ABC123',
+  submit: 'Join Class',
+  checking: 'Checking...',
+  invalidCode: 'Invalid code. Ask your teacher for the correct code.',
+  verifyError: 'Could not verify the code. Please check your connection and try again.',
+}
+
+export const TEACHER_LOGIN_COPY = {
+  title: 'Teacher Portal',
+  subtitle: 'Sign in to manage your modules',
+  signIn: 'Sign In',
+  createAccount: 'Create Account',
+  emailPlaceholder: 'Email address',
+  passwordPlaceholder: 'Password',
+  displayNamePlaceholder: 'Your name',
+  wait: 'Please wait...',
+  googleSignIn: 'Sign in with Google',
+  or: 'or',
+}
+
+export const CHAT_COPY = {
+  welcome:
+    "Hello! I'm Teacher's Pet, your math tutor. Feel free to ask me any math question — I'll walk you through it step by step.",
+  inputPlaceholder: 'Ask a math question...',
+  send: 'Send',
+  speechStart: 'Speak your question (microphone)',
+  speechStop: 'Stop dictation',
+  serverError:
+    "I couldn't reach the server. Make sure the backend is running on port 8000 and try again.",
+  reviewNotice: 'All messages and responses are sent directly to course staff for review.',
+  emptyStateHelp:
+    'I can help you with homework, explain tough concepts, or just practice.',
+  greetings: [
+    'Hey Math Wiz! What are we solving today?',
+    'Good morning, fellow math enthusiast!',
+    'Ready to crunch some numbers?',
+    'Hello, future mathematician!',
+    "Let's tackle some problems together!",
+    'Great to see you! Math awaits.',
+    'Hey there, problem solver!',
+    'Time to make math click!',
+    'Welcome! What shall we learn today?',
+    "Let's make math make sense!",
+  ],
+  quickActions: [
+    {
+      title: 'Homework Help',
+      subtitle: 'Walk me through a problem',
+      iconBg: 'bg-orange-100',
+      iconColor: 'text-orange-500',
+      response:
+        "Of course! Go ahead and type out your homework problem and I'll walk you through it step by step.",
+    },
+    {
+      title: 'Practice Problems',
+      subtitle: 'Give me something to solve',
+      iconBg: 'bg-blue-100',
+      iconColor: 'text-blue-500',
+      response:
+        "Let's practice! Tell me what topic you'd like to work on and I'll come up with a problem for you.",
+    },
+    {
+      title: 'Explain a Concept',
+      subtitle: 'Break it down simply',
+      iconBg: 'bg-purple-100',
+      iconColor: 'text-purple-500',
+      response: 'Happy to explain! What concept would you like me to break down for you?',
+    },
+    {
+      title: 'Study Tips',
+      subtitle: 'Help me prepare',
+      iconBg: 'bg-green-100',
+      iconColor: 'text-green-500',
+      response:
+        "Great thinking ahead! What subject or topic are you studying for? I'll share some tips to help you prepare.",
+    },
+  ],
+}
