+++ frontend/src/index.css
@@ -1,1 +1,2 @@
 @import "tailwindcss";
+@import "./theme/tokens.css";
