+++ frontend/src/theme/tokens.js
@@ -1,0 +1,49 @@
+export const themeTokens = {
+  color: {
+    bgCanvas: 'var(--color-bg-canvas)',
+    bgSurface: 'var(--color-bg-surface)',
+    bgMuted: 'var(--color-bg-muted)',
+    textPrimary: 'var(--color-text-primary)',
+    textSecondary: 'var(--color-text-secondary)',
+    textMuted: 'var(--color-text-muted)',
+    borderSubtle: 'var(--color-border-subtle)',
+    borderStrong: 'var(--color-border-strong)',
+    brand600: 'var(--color-brand-600)',
+    brand700: 'var(--color-brand-700)',
+    danger500: 'var(--color-danger-500)',
+    danger600: 'var(--color-danger-600)',
+  },
+  spacing: {
+    1: 'var(--space-1)',
+    2: 'var(--space-2)',
+    3: 'var(--space-3)',
+    4: 'var(--space-4)',
+    5: 'var(--space-5)',
+    6: 'var(--space-6)',
+    8: 'var(--space-8)',
+  },
+  radius: {
+    sm: 'var(--radius-sm)',
+    md: 'var(--radius-md)',
+    lg: 'var(--radius-lg)',
+    xl: 'var(--radius-xl)',
+    pill: 'var(--radius-pill)',
+  },
+  shadow: {
+    xs: 'var(--shadow-xs)',
+    sm: 'var(--shadow-sm)',
+    md: 'var(--shadow-md)',
+  },
+  motion: {
+    fast: 'var(--motion-fast)',
+    base: 'var(--motion-base)',
+    slow: 'var(--motion-slow)',
+    ease: 'var(--ease-standard)',
+  },
+  typography: {
+    sans: 'var(--font-sans)',
+    mono: 'var(--font-mono)',
+  },
+}
+
+export default themeTokens
